seperated_utils_functions
