+++ wsgi.py
@@ -1,7 +1,7 @@
 """App entry point."""
 from src import create_app
 
-app = create_app()
+app = create_app(test=False)
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=5000, debug=True)
